requirements.txt
django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
gunicorn==20.0.4
pyjwt==1.7.1